Add authorization to Reminder

+++ app/app/Http/Controllers/ReminderController.php
@@ -9,28 +9,40 @@ use App\Http\Resources\ReminderResource;
 use App\Models\Reminder;
 use App\Models\Vehicle;
 use Carbon\Carbon;
+use Illuminate\Support\Facades\Auth;
 use Symfony\Component\HttpFoundation\Response;
 
 class ReminderController extends Controller
 {
     public function index(Vehicle $vehicle)
     {
+        if ($vehicle->owner_id !== Auth::user()->id) {
+            return response()->json(data: ['message' => 'Forbidden'], status: Response::HTTP_FORBIDDEN);
+        }
+
         return new ReminderCollection($vehicle->reminders()->get());
     }
 
     public function show(Vehicle $vehicle, Reminder $reminder)
     {
+        if ($vehicle->owner_id !== Auth::user()->id) {
+            return response()->json(data: ['message' => 'Forbidden'], status: Response::HTTP_FORBIDDEN);
+        }
+
         return new ReminderResource($reminder);
     }
 
     public function store(Vehicle $vehicle, CreateReminderRequest $createReminderRequest)
     {
+        if ($vehicle->owner_id !== Auth::user()->id) {
+            return response()->json(data: ['message' => 'Forbidden'], status: Response::HTTP_FORBIDDEN);
+        }
+
         $input = $createReminderRequest->only(['name', 'last_remind', 'interval']);
         if (!empty($input['last_remind'])) {
             $input['next_remind'] = $this->updateNextRemind($input['last_remind'], $input['interval']);
         }
         $input['vehicle_id'] = $vehicle->id;
-
         Reminder::create($input);
 
         return response()->json(status: Response::HTTP_CREATED);
@@ -45,6 +57,10 @@ class ReminderController extends Controller
 
     public function update(Vehicle $vehicle, Reminder $reminder, EditReminderRequest $editReminderRequest)
     {
+        if ($vehicle->owner_id !== Auth::user()->id) {
+            return response()->json(data: ['message' => 'Forbidden'], status: Response::HTTP_FORBIDDEN);
+        }
+
         $input = $editReminderRequest->only(['name', 'last_remind', 'interval']);
         if (!empty($input['last_remind'])) {
             $input['next_remind'] = $this->updateNextRemind($input['last_remind'], $input['interval']);
@@ -59,6 +75,10 @@ class ReminderController extends Controller
 
     public function destroy(Vehicle $vehicle, Reminder $reminder)
     {
+        if ($vehicle->owner_id !== Auth::user()->id) {
+            return response()->json(data: ['message' => 'Forbidden'], status: Response::HTTP_FORBIDDEN);
+        }
+
         return response()->json(
             status: $reminder->delete() ?
             Response::HTTP_NO_CONTENT :
